+++ src/view/Camera.py
@@ -1,0 +1,1 @@
+class Camera:
