+++ src/view/UIManager.py
@@ -1,0 +1,1 @@
+class UIManager:
