+++ src/world/RoomManager.py
@@ -1,0 +1,4 @@
+
+class RoomManager:
+    def __init__(self):
+        self.rooms = {}
